Fixed number of CSS bugs (width/height, floats, etc). http://www.w3.org/Style/CSS/Test/current/sec5526c-test.htm ("acid test") looks quite good now.
Remaining problems: font (and em) size is wrong, background color is missing,
color of one text block is wrong and the inline form containing block
elements does not get rendered.

I have not clue what should be done with the last one. Lars?

svn path=/trunk/kdelibs/; revision=46554

+++ khtml/rendering/render_text.h
@@ -149,7 +149,7 @@ public:
 protected:
     short m_minWidth;
     short m_maxWidth;
-    short m_contentHeight;
+    int m_contentHeight;
 
     SelectionState m_selectionState : 3 ;
